Update to up to date.

+++ android/java/org/chromium/chrome/browser/crypto_wallet/util/AndroidUtils.java
@@ -84,6 +84,7 @@ public class AndroidUtils {
 
     /**
      * Gets device screen height in pixels, excluding the navigation bar (if visible) and insets.
+     *
      * @return device screen height in pixels.
      */
     public static int getScreenHeight() {
@@ -92,22 +93,27 @@ public class AndroidUtils {
 
     /**
      * Check if the fragment is safe to update its UI
+     *
      * @param frag instance
      * @return true if Fragment UI can be updated otherwise false
      */
     public static boolean canUpdateFragmentUi(Fragment frag) {
-        return !(frag.isRemoving() || frag.getActivity() == null || frag.isDetached()
-                || !frag.isAdded() || frag.getView() == null);
+        return !(frag.isRemoving()
+                || frag.getActivity() == null
+                || frag.isDetached()
+                || !frag.isAdded()
+                || frag.getView() == null);
     }
 
     /**
      * Calculated an ideal row count for shimmer effect based on screen size
+     *
      * @param skeletonRowHeight of a skeleton row view in pixels
      * @return count of rows for the skeleton list
      */
     public static int getSkeletonRowCount(int skeletonRowHeight) {
         int pxHeight = getScreenHeight();
-        return (int) Math.floor(pxHeight / skeletonRowHeight);
+        return (int) Math.floor((double) (pxHeight / ((double) skeletonRowHeight)));
     }
 
     /**
@@ -115,7 +121,7 @@ public class AndroidUtils {
      */
     public static boolean isDebugBuild() {
         return (ContextUtils.getApplicationContext().getApplicationInfo().flags
-                       & ApplicationInfo.FLAG_DEBUGGABLE)
+                        & ApplicationInfo.FLAG_DEBUGGABLE)
                 != 0;
     }
